chore: supprime le store legacy
