+++ src/pages/exploitation/parcellaire.vue
@@ -1,7 +1,6 @@
 <route lang="yaml">
 meta:
   requiresAuth: true
-  requiresGeodata: true
   seo:
     title: Gestion de mon parcellaire AB
 </route>
@@ -10,11 +9,11 @@ meta:
   <section class="fr-container fr-py-2w">
     <div class="fr-grid-row fr-grid-row--gutters">
       <div class="fr-col-6">
-        <OperatorSummary :operator="currentUser" :record="record" />
+        <OperatorSummary :operator="record.operator" :record="record" />
         <div class="fr-alert fr-alert--warning fr-mb-3w" v-if="hasFailures">
           <ValidationErrors :validation-result="validationResult" />
         </div>
-        <FeaturesTable :features="featureStore.collection" :operator="currentUser" :edit-form="EditForm" :validation-rules="validationRules" :mass-actions="massActions" />
+        <FeaturesTable :features="featureStore.collection" :operator="record.operator" :edit-form="EditForm" :validation-rules="validationRules" :mass-actions="massActions" />
       </div>
 
       <div class="fr-col-6 fr-col--map">
@@ -49,7 +48,7 @@ meta:
 </template>
 
 <script setup>
-import { computed, markRaw, readonly, ref, shallowRef, toRefs } from 'vue'
+import { computed, markRaw, readonly, ref, shallowRef } from 'vue'
 import bbox from '@turf/bbox'
 
 import FeaturesTable from '@/components/Features/Table.vue'
@@ -60,13 +59,15 @@ import EditForm from '@/components/Features/SingleItemOperatorForm.vue'
 import OperatorSummary from '@/components/Operator/Summary.vue'
 import CultureTypeModal from '@/components/Operator/CultureTypeModal.vue'
 
-import store from '@/store.js'
 import { useFeaturesStore } from '@/stores/features.js'
 import { applyValidationRules, OPERATOR_RULES } from '@/referentiels/ab.js'
 import surroundingsStyle from "@/map-styles/surroundings.json";
 import cadastreStyle from "@/map-styles/cadastre.json";
 import SurroundingsLegend from "@/components/Map/SurroundingsLegend.vue";
 import ValidationErrors from "@/components/Features/ValidationErrors.vue"
+import { getOperatorParcelles } from "@/cartobio-api.js"
+import { useUserStore } from "@/stores/index.js"
+import { useRouter } from "vue-router"
 
 const validationRules = readonly({
   rules: OPERATOR_RULES,
@@ -78,9 +79,24 @@ const massActions = [
   { label: 'Modifier les cultures', component: markRaw(CultureTypeModal) },
 ]
 
-const { currentUser, parcellaire, record } = toRefs(store.state)
+const router = useRouter()
+const userStore = useUserStore()
+let record
+try {
+  record = await getOperatorParcelles(userStore.user.id)
 
-const mapBounds = computed(() => bbox(parcellaire.value))
+  if (!record || !record.parcelles || !record.metadata.source) {
+    router.push('/exploitation/setup')
+  }
+} catch (error) {
+  // token has expired
+  if (error.response.status === 401) {
+    userStore.logout()
+    router.push('/login')
+  }
+}
+
+
 
 const showSurroundings = ref(false)
 
@@ -89,7 +105,9 @@ const isMapFullSize = ref(false)
 const mapRef = shallowRef(null)
 
 const featureStore = useFeaturesStore()
-featureStore.setAll(readonly(parcellaire.value.features))
+featureStore.setAll(readonly(record.parcelles.features))
+
+const mapBounds = computed(() => bbox(featureStore.collection))
 
 function setupWatchers (map) {
   mapRef.value = map
@@ -98,7 +116,7 @@ function setupWatchers (map) {
 }
 
 function focusOnOriginalBbox () {
-  zoomInto(parcellaire.value, { maxZoom: 12 })
+  zoomInto(featureStore.collection, { maxZoom: 12 })
 }
 
 function zoomInto (featureOrFeatureCollection, { maxZoom }) {
