+++ src/pages/exploitation/setup.vue
@@ -58,14 +58,12 @@ meta:
 </template>
 
 <script setup>
-import { ref, toRefs, readonly, computed } from 'vue'
+import { computed, readonly, ref } from 'vue'
 import { statsPush } from '@/stats.js'
-import store from '@/store.js'
 
 import OperatorSetup from '@/components/OperatorSetup/index.vue'
 
 const importTool = ref('')
-const { parcellaire } = toRefs(store.state)
 
 const currentStepIndex = ref(0)
 const allSteps = readonly([
@@ -85,7 +83,6 @@ function onSourceChange (source) {
 
 function onUploadStart () {
   statsPush(['trackEvent', 'setup', `import:${importTool.value}`, 'start'])
-  parcellaire.value = {}
   currentStepIndex.value = 1
 }
 
@@ -93,9 +90,8 @@ function onUploadPreview () {
   currentStepIndex.value += 1
 }
 
-function onSuccess ({ geojson }) {
+function onSuccess () {
   statsPush(['trackEvent', 'setup', `import:${importTool.value}`, 'ok'])
-  parcellaire.value = geojson
   currentStepIndex.value += 1
 }
